Restructure project to have a sln in the main folder
made a src\tinyOSCore folder
