+++ BuildAndRunAllTests.bat
@@ -1,7 +1,7 @@
 pushd .
-cd "OS Project"
+cd "src\TinyOSCore"
 dotnet build -c DEBUG
-copy "..\Sample Programs\*" bin\Debug
+copy "..\..\Sample Programs\*" bin\Debug
 pushd .
 cd bin\Debug
 call testAll.bat
