+++ BuildAndRunAllTests.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 set -v
 pushd .
-cd OS\ Project/
+cd src/TinyOSCore
 dotnet build -c Debug                              
-cp ../Sample\ Programs/* bin/Debug/                       
+cp ../../Sample\ Programs/* bin/Debug/                       
 pushd .                                                     
 cd bin/Debug/                                                
 ./testAll.sh
