--- OS Project/TinyOSCore.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26621.2
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TinyOSCore", "TinyOSCore.csproj", "{797096F0-9F9D-465E-80BE-ABED889E900E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{797096F0-9F9D-465E-80BE-ABED889E900E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{797096F0-9F9D-465E-80BE-ABED889E900E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{797096F0-9F9D-465E-80BE-ABED889E900E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{797096F0-9F9D-465E-80BE-ABED889E900E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {EC60F6D1-76C5-4C77-9BA8-D10F577809EC}
-	EndGlobalSection
-EndGlobal
