+++ src/TinyOSCore/appsettings.json
